+++ Libplanet/Blockchain/BlockChain.cs
@@ -629,7 +629,7 @@ namespace Libplanet.Blockchain
 
             foreach (Transaction<T> tx in stagedTransactions)
             {
-                if (!Policy.IsTransactionValid(tx))
+                if (!Policy.DoesTransactionFollowsPolicy(tx))
                 {
                     UnstageTransaction(tx);
                 }
@@ -761,9 +761,9 @@ namespace Libplanet.Blockchain
                 // the tx nounce order when the block was created
                 foreach (Transaction<T> tx1 in block.Transactions)
                 {
-                    if (!Policy.IsTransactionValid(tx1))
+                    if (!Policy.DoesTransactionFollowsPolicy(tx1))
                     {
-                        throw new InvalidTxByBlockPolicyException(
+                        throw new TxViolatingBlockPolicyException(
                             tx1.Id,
                             "According to BlockPolicy, this transaction is not valid.");
                     }