+++ Libplanet/Blockchain/Policies/IBlockPolicy.cs
@@ -20,12 +20,12 @@ namespace Libplanet.Blockchain.Policies
         IAction BlockAction { get; }
 
         /// <summary>
-        /// A predicate that determines if the transaction is valid.
+        /// A predicate that determines if the transaction follows the block policy.
         /// </summary>
-        /// <param name="transaction">A <see cref="Transaction{T}"/> to validate.</param>
+        /// <param name="transaction">A <see cref="Transaction{T}"/> to determine.</param>
         /// <returns><c>true</c> if <paramref name="transaction"/>is valid; otherwise, <c>false</c>.
         /// </returns>
-        bool IsTransactionValid(Transaction<T> transaction);
+        bool DoesTransactionFollowsPolicy(Transaction<T> transaction);
 
         /// <summary>
         /// Checks if <paramref name="nextBlock"/> is invalid, and if that